ColdEmailOutreach/requirements.txt
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
plotly
google-generativeai
mysql-connector-python
google-api-python-client